feat: custom themes (#61)
* feat: add `UserThemeRepository`

* feat: support custom themes in discord

* feat: allow selecting default themes

* feat: allow users to change their theme

* fix: improve theme menu

* feat: add theme edit menu

* feat: error handling

* fix: shadow configuration

* fix: allow turn labels in tyto notation

* feat: add line thickness parameter to HDS theme

* fix: improve `Color`

+++ web/src/jsMain/kotlin/de/mineking/hexo/web/Utils.kt
@@ -91,7 +91,7 @@ fun <T, U> State<T>.map(transform: (T) -> U) = object : State<U> {
 @Composable
 fun rememberTheme() = SettingsKey.Theme.collectAsState().map { it.theme }
 
-private val Color.css get() = rgba(red.toInt(), green.toInt(), blue.toInt(), alpha.toInt())
+private val Color.css get() = rgba(red, green, blue, alpha)
 fun BaseTheme.playerCssColor(owner: CellOwner) = playerColor(owner).css
 fun BaseTheme.playerColor(owner: CellOwner) = when (owner) {
     CellOwner.X -> playerXColor
